requirements.txt
streamlit
streamlit-audiorecorder
openai
gTTS
pydub
audioop-lts
